// src/domain/enums.ts
export enum GameOutcome {
  WIN = "WIN",
  LOSS = "LOSS",
  TIE = "TIE",
}

export enum GameType {
  REGULAR_SEASON = "REGULAR_SEASON",
  PLAYOFF = "PLAYOFF",
}

export enum PlayerPosition {
  QB = 0,
  RB = 2,
  WR = 4,
  TE = 6,
  DEFENSE = 16,
  KICKER = 17,
  BENCH = 20,
  IR = 21,
  FLEX = 23,
}